[Security Solution] Remove graphQL host (#94592)
* rm host from graphql

* clean up

* rm SortRequestDirection

* fix types

* rm comment

* fix types

* fix unit test

* fix unit test

* Update x-pack/plugins/security_solution/public/hosts/components/hosts_table/columns.tsx


* revert assertUnreachable

* use estypes

* rename details to HostDetails
