+++ x-pack/plugins/security_solution/common/search_strategy/security_solution/hosts/first_last_seen/index.ts
@@ -6,14 +6,14 @@
  */
 
 import { IEsSearchResponse } from '../../../../../../../../src/plugins/data/common';
-import { Inspect, Maybe } from '../../../common';
+import { Inspect, Maybe, Direction } from '../../../common';
 import { RequestOptionsPaginated } from '../..';
 import { HostsFields } from '../common';
 
 export interface HostFirstLastSeenRequestOptions
   extends Partial<RequestOptionsPaginated<HostsFields>> {
   hostName: string;
-  order: 'asc' | 'desc';
+  order: Direction.asc | Direction.desc;
 }
 
 export interface HostFirstLastSeenStrategyResponse extends IEsSearchResponse {
