+++ x-pack/plugins/security_solution/public/common/components/ml/influencers/host_to_influencers.ts
@@ -5,8 +5,8 @@
  * 2.0.
  */
 
+import { HostItem } from '../../../../../common/search_strategy/security_solution/hosts';
 import { InfluencerInput } from '../types';
-import { HostItem } from '../../../../graphql/types';
 
 export const hostToInfluencers = (hostItem: HostItem): InfluencerInput[] | null => {
   if (hostItem.host != null && hostItem.host.name != null) {
