+++ x-pack/plugins/security_solution/public/hosts/components/hosts_table/columns.tsx
@@ -78,7 +78,7 @@ export const getHostsColumns = (): HostsTableColumns => [
     hideForMobile: false,
     sortable: true,
     render: (lastSeen: Maybe<string | string[]> | undefined) => {
-      if (lastSeen != null) {
+      if (lastSeen != null && lastSeen.length > 0) {
         return (
           <FormattedRelativePreferenceDate
             value={Array.isArray(lastSeen) ? lastSeen[0] : lastSeen}
