+++ x-pack/plugins/security_solution/public/hosts/components/hosts_table/index.tsx
@@ -10,15 +10,6 @@ import { useDispatch } from 'react-redux';
 
 import { assertUnreachable } from '../../../../common/utility_types';
 import {
-  Direction,
-  HostFields,
-  HostItem,
-  HostsEdges,
-  HostsFields,
-  HostsSortField,
-  OsFields,
-} from '../../../graphql/types';
-import {
   Columns,
   Criteria,
   ItemsPerRow,
@@ -29,6 +20,14 @@ import { useDeepEqualSelector } from '../../../common/hooks/use_selector';
 import { hostsActions, hostsModel, hostsSelectors } from '../../store';
 import { getHostsColumns } from './columns';
 import * as i18n from './translations';
+import {
+  HostsEdges,
+  HostItem,
+  HostsSortField,
+  HostsFields,
+} from '../../../../common/search_strategy/security_solution/hosts';
+import { Direction } from '../../../../common/search_strategy';
+import { HostEcs, OsEcs } from '../../../../common/ecs/host';
 
 const tableType = hostsModel.HostsTableType.hosts;
 
@@ -45,10 +44,10 @@ interface HostsTableProps {
 }
 
 export type HostsTableColumns = [
-  Columns<HostFields['name']>,
+  Columns<HostEcs['name']>,
   Columns<HostItem['lastSeen']>,
-  Columns<OsFields['name']>,
-  Columns<OsFields['version']>
+  Columns<OsEcs['name']>,
+  Columns<OsEcs['version']>
 ];
 
 const rowItems: ItemsPerRow[] = [
@@ -82,7 +81,6 @@ const HostsTableComponent: React.FC<HostsTableProps> = ({
   const { activePage, direction, limit, sortField } = useDeepEqualSelector((state) =>
     getHostsSelector(state, type)
   );
-
   const updateLimitPagination = useCallback(
     (newLimit) =>
       dispatch(
@@ -178,6 +176,7 @@ const getNodeField = (field: HostsFields): string => {
   }
   assertUnreachable(field);
 };
+
 export const HostsTable = React.memo(HostsTableComponent);
 
 HostsTable.displayName = 'HostsTable';
