+++ x-pack/plugins/security_solution/public/hosts/pages/details/index.tsx
@@ -24,10 +24,8 @@ import { scoreIntervalToDateTime } from '../../../common/components/ml/score/sco
 import { SiemNavigation } from '../../../common/components/navigation';
 import { HostsDetailsKpiComponent } from '../../components/kpi_hosts';
 import { HostOverview } from '../../../overview/components/host_overview';
-import { manageQuery } from '../../../common/components/page/manage_query';
 import { SiemSearchBar } from '../../../common/components/search_bar';
 import { WrapperPage } from '../../../common/components/wrapper_page';
-import { HostOverviewByNameQuery } from '../../containers/hosts/details';
 import { useGlobalTime } from '../../../common/containers/use_global_time';
 import { useKibana } from '../../../common/lib/kibana';
 import { convertToBuildEsQuery } from '../../../common/lib/keury';
@@ -51,8 +49,7 @@ import { TimelineId } from '../../../../common/types/timeline';
 import { timelineDefaults } from '../../../timelines/store/timeline/defaults';
 import { useSourcererScope } from '../../../common/containers/sourcerer';
 import { useDeepEqualSelector, useShallowEqualSelector } from '../../../common/hooks/use_selector';
-
-const HostOverviewManage = manageQuery(HostOverview);
+import { useHostDetails } from '../../containers/hosts/details';
 
 const HostDetailsComponent: React.FC<HostDetailsProps> = ({ detailName, hostDetailsPagePath }) => {
   const dispatch = useDispatch();
@@ -96,6 +93,12 @@ const HostDetailsComponent: React.FC<HostDetailsProps> = ({ detailName, hostDeta
   );
 
   const { docValueFields, indicesExist, indexPattern, selectedPatterns } = useSourcererScope();
+  const [loading, { hostDetails: hostOverview, id }] = useHostDetails({
+    endDate: to,
+    startDate: from,
+    hostName: detailName,
+    indexNames: selectedPatterns,
+  });
   const filterQuery = convertToBuildEsQuery({
     config: esQuery.getEsQueryConfig(kibana.services.uiSettings),
     indexPattern,
@@ -131,49 +134,35 @@ const HostDetailsComponent: React.FC<HostDetailsProps> = ({ detailName, hostDeta
                 title={detailName}
               />
 
-              <HostOverviewByNameQuery
-                indexNames={selectedPatterns}
-                sourceId="default"
-                hostName={detailName}
-                skip={isInitializing}
+              <AnomalyTableProvider
+                criteriaFields={hostToCriteria(hostOverview)}
                 startDate={from}
                 endDate={to}
+                skip={isInitializing}
               >
-                {({ hostOverview, loading, id, inspect, refetch }) => (
-                  <AnomalyTableProvider
-                    criteriaFields={hostToCriteria(hostOverview)}
+                {({ isLoadingAnomaliesData, anomaliesData }) => (
+                  <HostOverview
+                    docValueFields={docValueFields}
+                    id={id}
+                    isInDetailsSidePanel={false}
+                    data={hostOverview as HostItem}
+                    anomaliesData={anomaliesData}
+                    isLoadingAnomaliesData={isLoadingAnomaliesData}
+                    indexNames={selectedPatterns}
+                    loading={loading}
                     startDate={from}
                     endDate={to}
-                    skip={isInitializing}
-                  >
-                    {({ isLoadingAnomaliesData, anomaliesData }) => (
-                      <HostOverviewManage
-                        docValueFields={docValueFields}
-                        id={id}
-                        inspect={inspect}
-                        isInDetailsSidePanel={false}
-                        refetch={refetch}
-                        setQuery={setQuery}
-                        data={hostOverview as HostItem}
-                        anomaliesData={anomaliesData}
-                        isLoadingAnomaliesData={isLoadingAnomaliesData}
-                        indexNames={selectedPatterns}
-                        loading={loading}
-                        startDate={from}
-                        endDate={to}
-                        narrowDateRange={(score, interval) => {
-                          const fromTo = scoreIntervalToDateTime(score, interval);
-                          setAbsoluteRangeDatePicker({
-                            id: 'global',
-                            from: fromTo.from,
-                            to: fromTo.to,
-                          });
-                        }}
-                      />
-                    )}
-                  </AnomalyTableProvider>
+                    narrowDateRange={(score, interval) => {
+                      const fromTo = scoreIntervalToDateTime(score, interval);
+                      setAbsoluteRangeDatePicker({
+                        id: 'global',
+                        from: fromTo.from,
+                        to: fromTo.to,
+                      });
+                    }}
+                  />
                 )}
-              </HostOverviewByNameQuery>
+              </AnomalyTableProvider>
 
               <EuiHorizontalRule />
