+++ x-pack/plugins/security_solution/public/hosts/store/model.ts
@@ -5,7 +5,8 @@
  * 2.0.
  */
 
-import { Direction, HostsFields } from '../../graphql/types';
+import { HostsFields } from '../../../common/search_strategy/security_solution/hosts';
+import { Direction } from '../../graphql/types';
 
 export enum HostsType {
   page = 'page',
