+++ x-pack/plugins/security_solution/public/hosts/store/reducer.ts
@@ -6,8 +6,8 @@
  */
 
 import { reducerWithInitialState } from 'typescript-fsa-reducers';
+import { Direction, HostsFields } from '../../../common/search_strategy';
 
-import { Direction, HostsFields } from '../../graphql/types';
 import { DEFAULT_TABLE_ACTIVE_PAGE, DEFAULT_TABLE_LIMIT } from '../../common/store/constants';
 
 import {
