+++ x-pack/plugins/security_solution/public/overview/components/host_overview/endpoint_overview/index.tsx
@@ -13,9 +13,11 @@ import { OverviewDescriptionList } from '../../../../common/components/overview_
 import { DescriptionList } from '../../../../../common/utility_types';
 import { getEmptyTagValue } from '../../../../common/components/empty_value';
 import { DefaultFieldRenderer } from '../../../../timelines/components/field_renderers/field_renderers';
-import { EndpointFields, HostPolicyResponseActionStatus } from '../../../../graphql/types';
-
 import * as i18n from './translations';
+import {
+  EndpointFields,
+  HostPolicyResponseActionStatus,
+} from '../../../../../common/search_strategy/security_solution/hosts';
 
 interface Props {
   contextID?: string;
