+++ x-pack/plugins/security_solution/public/timelines/components/side_panel/host_details/expandable_host.test.tsx
@@ -70,9 +70,7 @@ describe('Expandable Host Component', () => {
         </TestProviders>
       );
 
-      expect(wrapper.find('HostOverviewByNameComponentQuery').prop('indexNames')).toStrictEqual([
-        'IShouldBeUsed',
-      ]);
+      expect(wrapper.find('HostOverview').prop('indexNames')).toStrictEqual(['IShouldBeUsed']);
     });
   });
 });
