+++ x-pack/plugins/security_solution/server/lib/types.ts
@@ -10,7 +10,6 @@ export { ConfigType as Configuration } from '../config';
 import type { SecuritySolutionRequestHandlerContext } from '../types';
 
 import { FrameworkAdapter, FrameworkRequest } from './framework';
-import { Hosts } from './hosts';
 import { IndexFields } from './index_fields';
 import { SourceStatus } from './source_status';
 import { Sources } from './sources';
@@ -19,11 +18,8 @@ import { PinnedEvent } from './timeline/saved_object/pinned_events';
 import { Timeline } from './timeline/saved_object/timelines';
 import { TotalValue, BaseHit, Explanation } from '../../common/detection_engine/types';
 
-export * from './hosts';
-
 export interface AppDomainLibs {
   fields: IndexFields;
-  hosts: Hosts;
 }
 
 export interface AppBackendLibs extends AppDomainLibs {
@@ -142,54 +138,9 @@ export interface Hits<T, U> {
     hits: U[];
   };
 }
-export type SortRequestDirection = 'asc' | 'desc';
-
-interface SortRequestField {
-  [field: string]: SortRequestDirection;
-}
-
-export type SortRequest = SortRequestField[];
 
 export interface MSearchHeader {
   index: string[] | string;
   allowNoIndices?: boolean;
   ignoreUnavailable?: boolean;
-}
-
-export interface AggregationRequest {
-  [aggField: string]: {
-    terms?: {
-      field?: string;
-      missing?: string;
-      size?: number;
-      script?: {
-        source: string;
-        lang: string;
-      };
-      order?: {
-        [aggSortField: string]: SortRequestDirection;
-      };
-    };
-    max?: {
-      field: string;
-    };
-    aggs?: {
-      [aggSortField: string]: {
-        [aggType: string]: {
-          field: string;
-        };
-      };
-    };
-    top_hits?: {
-      size?: number;
-      sort?: Array<{
-        [aggSortField: string]: {
-          order: SortRequestDirection;
-        };
-      }>;
-      _source: {
-        includes: string[];
-      };
-    };
-  };
 }
