+++ x-pack/plugins/security_solution/server/search_strategy/security_solution/factory/hosts/all/helpers.ts
@@ -8,9 +8,13 @@
 import { set } from '@elastic/safer-lodash-set/fp';
 import { get, has, head } from 'lodash/fp';
 import { hostFieldsMap } from '../../../../../../common/ecs/ecs_fields';
-import { HostsEdges } from '../../../../../../common/search_strategy/security_solution/hosts';
+import {
+  HostAggEsItem,
+  HostBuckets,
+  HostsEdges,
+  HostValue,
+} from '../../../../../../common/search_strategy/security_solution/hosts';
 
-import { HostAggEsItem, HostBuckets, HostValue } from '../../../../../lib/hosts/types';
 import { toObjectArrayOfStrings } from '../../../../helpers/to_array';
 
 export const HOSTS_FIELDS: readonly string[] = [
