+++ x-pack/plugins/security_solution/server/search_strategy/security_solution/factory/hosts/details/query.host_details.dsl.ts
@@ -8,9 +8,8 @@
 import { ISearchRequestParams } from '../../../../../../../../../src/plugins/data/common';
 import { cloudFieldsMap, hostFieldsMap } from '../../../../../../common/ecs/ecs_fields';
 import { HostDetailsRequestOptions } from '../../../../../../common/search_strategy/security_solution';
-import { buildFieldsTermAggregation } from '../../../../../lib/hosts/helpers';
 import { reduceFields } from '../../../../../utils/build_query/reduce_fields';
-import { HOST_FIELDS } from './helpers';
+import { HOST_FIELDS, buildFieldsTermAggregation } from './helpers';
 
 export const buildHostDetailsQuery = ({
   hostName,
