+++ x-pack/plugins/security_solution/server/search_strategy/security_solution/factory/hosts/last_first_seen/__mocks__/index.ts
@@ -6,6 +6,7 @@
  */
 
 import {
+  Direction,
   HostFirstLastSeenRequestOptions,
   HostsQueries,
 } from '../../../../../../../common/search_strategy';
@@ -23,7 +24,7 @@ export const mockOptions: HostFirstLastSeenRequestOptions = {
   docValueFields: [],
   factoryQueryType: HostsQueries.firstOrLastSeen,
   hostName: 'siem-kibana',
-  order: 'asc',
+  order: Direction.asc,
 };
 
 export const mockSearchStrategyFirstSeenResponse = {
@@ -141,7 +142,7 @@ export const formattedSearchStrategyFirstResponse = {
             sort: [
               {
                 '@timestamp': {
-                  order: 'asc',
+                  order: Direction.asc,
                 },
               },
             ],
@@ -206,7 +207,7 @@ export const formattedSearchStrategyLastResponse = {
             sort: [
               {
                 '@timestamp': {
-                  order: 'desc',
+                  order: Direction.desc,
                 },
               },
             ],
@@ -237,6 +238,6 @@ export const expectedDsl = {
     _source: ['@timestamp'],
     query: { bool: { filter: [{ term: { 'host.name': 'siem-kibana' } }] } },
     size: 1,
-    sort: [{ '@timestamp': { order: 'asc' } }],
+    sort: [{ '@timestamp': { order: Direction.asc } }],
   },
 };
